requirements-server.txt
# requirements solo para el servidor (sin Kivy, más liviano para deploy)
pillow>=10.0.0
requests>=2.31.0

# ML — inferencia
torch>=2.0.0
torchvision>=0.15.0

# Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
